Simplify lazeresolver
