+++ cms/urls.py
@@ -1,9 +1,10 @@
+from functools import cached_property
+
 from django.conf import settings
-from django.urls import include, path, re_path
+from django.urls import URLResolver, include, path, re_path
 from django.urls.resolvers import RegexPattern
 
 from cms import views
-from cms.appresolver import LazyURLResolver
 from cms.constants import SLUG_REGEXP
 
 if settings.APPEND_SLASH:
@@ -22,6 +23,22 @@ def _get_apphook_urlpatterns():
     from cms.appresolver import get_app_patterns
 
     return get_app_patterns()
+
+
+class LazyURLResolver(URLResolver):
+    def __init__(self, *args, get_urlpatterns, **kwargs):
+        self._get_urlpatterns = get_urlpatterns
+        super().__init__(*args, **kwargs)
+
+    @property
+    def urlconf_module(self):
+        # Django allows urlconf_module to be a list of URLPattern/URLResolver.
+        return self.url_patterns
+
+    @cached_property
+    def url_patterns(self):
+        patterns = self._get_urlpatterns() or []
+        return list(patterns)
 
 
 urlpatterns = [
